split qlinear concat and qdq families
