+++ onnx2tf/tflite_builder/op_builders/__init__.py
@@ -172,9 +172,10 @@ from onnx2tf.tflite_builder.op_builders.control import (
 )
 from onnx2tf.tflite_builder.op_builders.quantized import (
     build_conv_integer_op,
-    build_dequantize_linear_op,
-    build_qlinear_concat_op,
     build_qlinear_conv_op,
+)
+from onnx2tf.tflite_builder.op_builders.quantize_linear import (
+    build_dequantize_linear_op,
     build_quantize_linear_op,
 )
 from onnx2tf.tflite_builder.op_builders.dynamic_quantize import (
@@ -187,6 +188,9 @@ from onnx2tf.tflite_builder.op_builders.qlinear_fc import (
 from onnx2tf.tflite_builder.op_builders.qlinear_binary import (
     build_qlinear_add_op,
     build_qlinear_mul_op,
+)
+from onnx2tf.tflite_builder.op_builders.qlinear_concat import (
+    build_qlinear_concat_op,
 )
 from onnx2tf.tflite_builder.op_builders.qlinear_activation import (
     build_qlinear_leaky_relu_op,
